retail-etl-td-to-bigquery/requirements.txt
# Python 3.11+
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.0.0
teradatasql>=17.0.0
pyarrow>=12.0.0
pandas>=2.0.0
